pyproject.toml
[tool.poetry]
name = "advent_of_code_2022"
version = "0.1.0"
description = ""
authors = ["Matěj Račinský <racinsky.matej@seznam.cz>"]

[tool.poetry.dependencies]
python = "^3.9"
aocd = "^0.1"
psycopg2 = "^2.9.5"
PyYAML = "^6.0"
numpy = "^1.23.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
